issue #2 rename data to dataset
