+++ deepreg/dataset/load.py
@@ -1,10 +1,10 @@
 import os
 
-from deepreg.data.loader.grouped_loader import GroupedDataLoader
-from deepreg.data.loader.h5_loader import H5FileLoader
-from deepreg.data.loader.nifti_loader import NiftiFileLoader
-from deepreg.data.loader.paired_loader import PairedDataLoader
-from deepreg.data.loader.unpaired_loader import UnpairedDataLoader
+from deepreg.dataset.loader.grouped_loader import GroupedDataLoader
+from deepreg.dataset.loader.h5_loader import H5FileLoader
+from deepreg.dataset.loader.nifti_loader import NiftiFileLoader
+from deepreg.dataset.loader.paired_loader import PairedDataLoader
+from deepreg.dataset.loader.unpaired_loader import UnpairedDataLoader
 
 
 def get_data_loader(data_config, mode):
